fix: remove `generateRoutesFromFiles` from runtime code
